ADD BASIC GAME CONTROLLER LOGIC
I added some of the logic behind the controller. There are some dependencies due to the fact that the board and view aren't finished yet. However, the click on the column should trigger the appropriate actions. Beyond the controller itself, the buttons should have an id in order to make this simpler and the layout should load the controller AFTER the board.js and view.js have loaded. Otherwise, their prototypes wouldn't be available.